requirements.txt
flet
rembg
Pillow